pyproject.toml
[project]
name = "langgraph-shopping-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.40",
    "langchain-google-genai>=2.0.11",
    "langgraph>=0.2.75",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
